requirements.txt
certifi==2017.4.17
chardet==3.0.3
click==6.7
Flask==0.12.2
google-api-python-client==1.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauth2client==4.1.0
oauthlib==2.0.2
psutil==5.2.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
python-gnupg==0.4.0
requests==2.17.3
requests-oauth2==0.2.0
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
Werkzeug==0.12.2
